requirements.txt
torch~=2.7.0
torchvision~=0.22.0
Flask~=3.1.0
Werkzeug~=3.1.3
pillow~=11.2.1
gunicorn~=23.0.0